rename d_exit->d_quit
