+++ Source/d_quit.c
